Split type context into generic context and instantiation
