+++ lang/syntax/src/ctx/mod.rs
@@ -1,5 +1,7 @@
+pub mod context;
 mod def;
 mod levels;
+
 pub mod map_idx;
 pub mod values;
